requirements-api.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch<2.2.0
torchvision<0.17.0
fastai<2.7.14
fastcore<1.5.30
plum-dispatch==1.7.4
fasttransform
cloudpickle
numpy<2
fastapi
uvicorn
python-multipart
sqlalchemy>=2.0
psycopg2-binary
pydantic-settings
ipython
